+++ haystack/nodes/preprocessor/merger.py
@@ -1,13 +1,29 @@
-from typing import Optional, List, Dict, Any, Union, Tuple
+from typing import List, Optional, Tuple, Union, Dict, Any, Callable
+
+try:
+    from typing import Literal
+except ImportError:
+    from typing_extensions import Literal  # type: ignore
 
 import logging
+from pathlib import Path
 from copy import deepcopy
-from math import inf
 
 from tqdm import tqdm
+from transformers import PreTrainedTokenizer
 
 from haystack.schema import Document
 from haystack.nodes.base import BaseComponent
+from haystack.modeling.model.feature_extraction import FeatureExtractor
+from haystack.nodes.preprocessor.helpers import (
+    split_by_separators,
+    split_by_transformers_tokenizer,
+    validate_unit_boundaries,
+    make_merge_groups,
+    merge_headlines,
+    common_values,
+)
+
 
 logger = logging.getLogger(__name__)
 
@@ -39,6 +55,9 @@ class DocumentMerger(BaseComponent):
         realign_headlines: bool = True,
         retain_page_number: bool = True,
         progress_bar: bool = True,
+        tokenizer_model: Optional[
+            Union[Literal["word"], Path, PreTrainedTokenizer, FeatureExtractor, Callable]
+        ] = "word",
         id_hash_keys: Optional[List[str]] = None,
     ):
         """
@@ -112,6 +131,18 @@ class DocumentMerger(BaseComponent):
                             add the number of tokens it contains to each source document's `tokens_count` metadata field
                             to compensate.
 
+        :param tokenizer_model: If `max_tokens` is set, you should provide a tokenizer model to compute the tokens.
+                                There are several options, depending on the tradeoff you need between precision and speed:
+                                - "word". The text is split with the `split()` function.
+                                - A tokenizer model. You can give its identifier on Hugging Face Hub, a local path to load it from,
+                                    or an instance of `FeatureExtractor` or Transformer's `PreTrainedTokenizer`.
+                                - A lambda function. In this case, make sure it takes one single input parameter called `text`, like
+                                  `tokenizer_model=lambda text: text.split("my token delimiter")`
+
+                                Note that a tokenizer is not mandatory as you can also provide the tokens manually to the `merge` method.
+
+                                Defaults to "word".
+
         :param id_hash_keys: the value to pass to id_hash_keys when initializing the merged Document(s).
         """
         super().__init__()
@@ -127,6 +158,37 @@ class DocumentMerger(BaseComponent):
         self.retain_page_number = retain_page_number
         self.progress_bar = progress_bar
         self.id_hash_keys = id_hash_keys
+
+        self._tokenizer = None
+        if tokenizer_model or max_tokens:
+            self.tokenizer = tokenizer_model
+
+    @property
+    def tokenizer(self):
+        return self._tokenizer
+
+    @tokenizer.setter
+    def tokenizer(self, tokenizer_model=Union[str, Path, PreTrainedTokenizer, FeatureExtractor, Callable]):
+        if not tokenizer_model:
+            raise ValueError(
+                "Can't set the tokenizer to None. "
+                "Provide either the string 'word', a Hugging Face identifier, a path to a local tokenizer, "
+                "or an instance of Haystack's FeatureExtractor or Transformers' PreTrainedTokenizer. "
+                "You can also provide your own lambda function to tokenize text: in this case "
+                "make sure it takes only one input parameter called 'text'."
+            )
+        if isinstance(tokenizer_model, (PreTrainedTokenizer, FeatureExtractor)):
+            self._tokenizer = lambda text: split_by_transformers_tokenizer(text=text, tokenizer=tokenizer_model)[0]
+
+        elif isinstance(tokenizer_model, (str, Path)):
+            if tokenizer_model == "word":
+                self._tokenizer = lambda text: split_by_separators(text=text, separators=[" ", "\n", "\f"])
+            else:
+                self._tokenizer = lambda text: split_by_transformers_tokenizer(
+                    text=text, tokenizer=FeatureExtractor(pretrained_model_name_or_path=tokenizer_model)
+                )[0]
+        else:
+            self._tokenizer = tokenizer_model
 
     def _validate_window_params(self, window_size: int, window_overlap: int):
         """
@@ -230,6 +292,13 @@ class DocumentMerger(BaseComponent):
 
         :param id_hash_keys: the value to pass to id_hash_keys when initializing the merged Document(s).
         """
+        max_tokens = max_tokens if max_tokens is not None else self.max_tokens
+
+        tokens = []
+        if max_tokens:
+            for document in documents:
+                tokens += self.tokenizer(document.content)
+
         merged_documents = self.merge(
             documents=documents,
             separator=separator,
@@ -241,6 +310,7 @@ class DocumentMerger(BaseComponent):
             max_tokens=max_tokens,
             max_chars=max_chars,
             id_hash_keys=id_hash_keys,
+            tokens=tokens,
         )
         return {"documents": merged_documents}, "output_1"
 
@@ -436,6 +506,15 @@ class DocumentMerger(BaseComponent):
                             add the number of tokens it contains to each source document's `tokens_count` metadata field
                             to compensate.
 
+        :param tokens: a single list with all the tokens contained in all documents to merge.
+                        So, if the document look like `[Document(content="hello how are you"), Document(content="I'm fine thanks")]`,
+                        `tokens` should contain something similar to `["hello", "how", "are", "you", "I'm", "fine", "thanks"]`.
+
+                        If the `tokens` are not given but `max_tokens` is set, `DocumentMerger` will tokenize the text by itself
+                        before proceeding.
+
+                        Useful to speed up the token counting process when `max_tokens` is set. Ignored if `max_tokens` is not set.
+
         :param id_hash_keys: the value to pass to id_hash_keys when initializing the merged Document(s).
         """
         if not documents:
@@ -459,10 +538,18 @@ class DocumentMerger(BaseComponent):
         id_hash_keys = id_hash_keys if id_hash_keys is not None else self.id_hash_keys
         self._validate_window_params(window_size=window_size, window_overlap=window_overlap)
 
+        if max_tokens and not tokens:
+            logger.info(
+                "DocumentMerger.merge() received `max_tokens` but no `tokens`. Generating the tokens on the fly."
+            )
+            tokens = []
+            for document in documents:
+                tokens += self.tokenizer(document.content)
+
         valid_contents = validate_unit_boundaries(
             contents=[doc.content for doc in documents], max_chars=max_chars, max_tokens=max_tokens, tokens=tokens
         )
-        groups_to_merge = make_groups(
+        groups_to_merge = make_merge_groups(
             contents=valid_contents,
             window_size=window_size,
             window_overlap=window_overlap,
@@ -501,243 +588,3 @@ class DocumentMerger(BaseComponent):
                     merged_documents[group_index].meta["page"] = min(int(page) for page in pages)
 
         return merged_documents
-
-
-def make_groups(
-    contents: List[Tuple[str, int]], window_size: int, window_overlap: int, max_tokens: int, max_chars: int
-):
-    """
-    Creates the groups of documents that need to be merged, respecting all boundaries.
-
-    :param contents: a tuple with the document content and how many tokens it contains.
-        Can be created with `validate_boundaries()`
-    :param window_size: The number of documents to include in each merged batch. For example,
-        if set to 2, the documents are merged in pairs. When set to 0, merges all documents
-        into one single document.
-    :param window_overlap: Applies a sliding window approach over the documents groups.
-        For example, if `window_size=3` and `window_overlap=2`, the resulting documents come
-        from the merge of the following groups: `[doc1, doc2, doc3]`, `[doc2, doc3, doc4]`, ...
-    :param max_tokens: the maximum number of tokens allowed
-    :param max_chars: the maximum number of chars allowed
-    :returns: a list of lists, each sublist containing the indices of the documents that should be merged together.
-        The length of this list is equal to the number of documents that will be produced by the merger.
-    """
-    groups = []
-    if max_tokens:
-        group: List[int] = []
-        content_len = 0
-        tokens_count = 0
-        doc_index = 0
-        while doc_index < len(contents):
-
-            if max_chars and content_len + len(contents[doc_index][0]) > max_chars:
-                # Rare and odd case: log loud
-                logger.warning(
-                    "One document reached `max_chars` (%s) before either `max_tokens` (%s) or `window_size` (%s). "
-                    "The last unit is moved to the next document to keep the chars count below the threshold."
-                    "Consider raising `max_chars` and double-check how the input coming from earlier nodes looks like.\n"
-                    "Enable DEBUG logs to see the document.",
-                    max_chars,
-                    max_tokens,
-                    window_size,
-                )
-                logger.debug(
-                    "********* Original document content: *********\n%s\n****************************",
-                    contents[doc_index][0],
-                )
-                groups.append(group)
-                group = []
-                tokens_count = 0
-                content_len = 0
-                if window_overlap:
-                    doc_index -= window_overlap
-
-            if max_tokens and tokens_count + contents[doc_index][1] > max_tokens:
-                # More plausible case: log, but not too loud
-                logger.info(
-                    "One document reached `max_tokens` (%s) before `window_size` (%s). "
-                    "The last unit is moved to the next document to keep the token count below the threshold.\n"
-                    "Enable DEBUG logs to see the document.",
-                    max_tokens,
-                    window_size,
-                )
-                logger.debug(
-                    "********* Original document content: *********\n%s\n****************************",
-                    contents[doc_index][0],
-                )
-                groups.append(group)
-                group = []
-                tokens_count = 0
-                content_len = 0
-                if window_overlap:
-                    doc_index -= window_overlap
-
-            if window_size and len(group) >= window_size:
-                # Fully normal: debug log only
-                logger.debug(
-                    "One document reached `window_size` (%s) before `max_tokens` (%s). ", max_tokens, window_size
-                )
-                groups.append(group)
-                group = []
-                tokens_count = 0
-                content_len = 0
-                if window_overlap:
-                    doc_index -= window_overlap
-
-            # Still accumulating
-            group.append(doc_index)
-            tokens_count += contents[doc_index][1]
-            content_len += len(contents[doc_index][0])
-            doc_index += 1
-
-        # Last group after the loop
-        if group:
-            group.append(doc_index)
-        return groups
-
-    # Shortcuts for when max_tokens is not used
-    elif window_size:
-        return [
-            list(range(pos, pos + window_size))
-            for pos in range(0, max(1, len(contents) - window_overlap), window_size - window_overlap)
-        ]
-    else:
-        return [list(range(len(contents)))]
-
-
-def validate_unit_boundaries(
-    contents: List[str], max_chars: int, max_tokens: int, tokens: Optional[List[str]] = None
-) -> List[Tuple[str, int]]:
-    """
-    Makes sure all boundaries (max_tokens if given, max_char if given) are respected. Splits the strings if necessary.
-
-    :param contents: the content of all documents to merge, as a string
-    :param tokens: a single list with all the tokens contained in all documents to merge.
-        So, if `contents = ["hello how are you", "I'm fine thanks"]`,
-        tokens should contain something similar to `["hello", "how", "are", "you", "I'm", "fine", "thanks"]`
-    :param max_tokens: the maximum amount of tokens allowed in a doc
-    :param max_chars: the maximum number of chars allowed in a doc
-    :return: a tuple (content, n_of tokens) if max_token is set, else (content, 0)
-    """
-    valid_contents = []
-
-    # Count tokens and chars, split if necessary
-    if max_tokens:
-        if not tokens:
-            raise ValueError("if max_tokens is set, you must pass the tokenized text to `tokens`.")
-
-        for content in contents:
-            tokens_length = 0
-            for tokens_count, token in enumerate(tokens):
-                tokens_length += len(token)
-
-                # If we reached the doc length, record how many tokens it contained and pass on the next doc
-                if tokens_length >= len(content):
-                    valid_contents.append((content, tokens_count))
-                    break
-
-                # This doc has more than max_tokens: save the head as a separate document and continue
-                if tokens_count >= max_tokens:
-                    logger.info(
-                        "Found unit of text with a token count higher than the maximum allowed. "
-                        "The unit is going to be cut at %s tokens, and the remaining %s chars will go to one (or more) new documents. "
-                        "Set the maximum amout of tokens allowed through the 'max_tokens' parameter."
-                        "Keep in mind that very long Documents can severely impact the performance of Readers.\n"
-                        "Enable DEBUG level logs to see the content of the unit that is being split",
-                        max_tokens,
-                        len(content) - tokens_length,
-                    )
-                    logger.debug(
-                        "********* Original document content: *********\n%s\n****************************", content
-                    )
-                    valid_contents.append((content[:tokens_length], tokens_count))
-                    content = content[:tokens_length]
-                    tokens_count = 0
-
-                # This doc has more than max_chars: save the head as a separate document and continue
-                if max_chars and tokens_length >= max_chars:
-                    logger.warning(
-                        "Found unit of text with a character count higher than the maximum allowed. "
-                        "The unit is going to be cut at %s chars, so %s chars are being moved to one (or more) new documents. "
-                        "Set the maximum amout of characters allowed through the 'max_chars' parameter. "
-                        "Keep in mind that very long Documents can severely impact the performance of Readers.\n"
-                        "Enable DEBUG level logs to see the content of the unit that is being split",
-                        max_chars,
-                        len(content) - max_chars,
-                    )
-                    logger.debug(
-                        "********* Original document content: *********\n%s\n****************************", content
-                    )
-                    valid_contents.append((content[:max_chars], tokens_count))
-                    content = content[:max_chars]
-                    tokens_count = 0
-
-    # Validate only the chars, split if necessary
-    else:
-        for content in contents:
-            if max_chars and len(content) >= max_chars:
-                logger.error(
-                    "Found unit of text with a character count higher than the maximum allowed. "
-                    "The unit is going to be cut at %s chars, so %s chars are being moved to one (or more) new documents. "
-                    "Set the maximum amout of characters allowed through the 'max_chars' parameter. "
-                    "Keep in mind that very long Documents can severely impact the performance of Readers.",
-                    max_chars,
-                    len(content) - max_chars,
-                )
-                valid_contents += [
-                    (content[max_chars * i : max_chars * (i + 1)], 0) for i in range(int(len(content) / max_chars) + 1)
-                ]
-            else:
-                valid_contents.append((content, 0))
-
-    return valid_contents
-
-
-def merge_headlines(
-    sources: List[Tuple[str, List[Dict[str, Any]]]], separator: str
-) -> List[Dict[str, Union[str, int]]]:
-    """
-    Merges the headlines dictionary with the new position of each headline into the merged document.
-    Assumes the documents are in the same order as when they were merged.
-
-    :param sources: tuple (source document content, source document headlines).
-    :param separator: the string used to join the document's content
-    :return: a dictionary that can be assigned to the merged document's headlines key.
-    """
-    aligned_headlines = []
-    position_in_merged_document = 0
-    for content, headlines in sources:
-        for headline in headlines:
-            headline["start_idx"] += position_in_merged_document
-            aligned_headlines.append(headline)
-        position_in_merged_document += len(content) + len(separator)
-    return aligned_headlines
-
-
-def common_values(list_of_dicts: List[Dict[str, Any]], exclude: List[str]) -> Dict[str, Any]:
-    """
-    Retains all keys shared across all the documents being merged.
-
-    Such keys are checked recursively, see tests.
-
-    :param list_of_dicts: dicts to merge
-    :param exclude: keys to drop regardless of their content
-    :return: the merged dictionary
-    """
-    shortest_dict = min(list_of_dicts, key=len)
-    merge_dictionary = {}
-    for key, value in shortest_dict.items():
-
-        # if not all dicts have this key, skip
-        if not key in exclude and all(key in dict.keys() for dict in list_of_dicts):
-
-            # if the value is a dictionary, merge recursively
-            if isinstance(value, dict):
-                list_of_subdicts = [dictionary[key] for dictionary in list_of_dicts]
-                merge_dictionary[key] = common_values(list_of_subdicts, exclude=[])
-
-            # If the value is not a dictionary, keep only if the values is the same for all
-            elif all(value == dict[key] for dict in list_of_dicts):
-                merge_dictionary[key] = value
-
-    return merge_dictionary
